refactor(server): split account relations and register deletedAt as a better-auth additional field
\`accounts.ts\` is intended to stay close to what
\`@better-auth/cli generate\` emits so we can regenerate it as the
better-auth schema evolves. We had two things working against that:

1. The custom \`deletedAt\` column on \`user\` was not declared anywhere
   on the better-auth side, so a regeneration would silently drop it
   (or, worse, silently succeed and break soft-delete). It also forced
   us to cast \`session.user\` whenever we wanted to read it.
2. Drizzle \`relations(...)\` were inlined alongside the table
   definitions, which the CLI would clobber on regenerate.

- Move every \`relations(...)\` export into a sibling
  \`accounts-extensions.ts\` and re-export it from \`schemas/index.ts\`
  so consumers see no behavior change.
- Trim \`accounts.ts\` down to pure \`pgTable\` defs with a header
  comment explaining the regeneration contract and a NOTICE on
  \`deletedAt\` pointing back at the better-auth declaration.
- Declare \`deletedAt\` in
  \`createAuth().user.additionalFields\` (\`type: 'date'\`,
  \`input: false\`, \`returned: false\`, \`fieldName: 'deleted_at'\`).
  This pins the column on the better-auth side, ensures the next
  CLI generation keeps it, and keeps it private from public APIs.

Made-with: Cursor
